requirements.txt
Flask==2.3.3
openai==2.6.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==20.1.0
Werkzeug==2.3.7
Pillow==12.0.0

